separate grids,  making separate sounds, but still having trouble playing more than one note at a time
